// lib/types/types.ts
export interface UserProfile {
  uid: string;
  name: string;
  birthDate: string;
  birthTime: string;
  birthLocation: string;
  gender: string;
  onboardingCompleted: boolean;
}

export interface ChildChart {
  id: string;
  name: string;
  birthDate: string;
  birthTime: string;
  birthLocation: string;
}